rm extra file
